fix(sprint): preserve cached preflight on resume
